fix: 优化了历史进度定位的逻辑，修复了点击已失效视频详情会崩落的问题，修复了一些因为int/long引起的问题，修复了转发动态无法点击的问题，修复了快速定位到动态页时标题为空的问题，修复了加载视频时没有小电视的问题

+++ app/src/main/java/cn/spacexc/wearbili/dataclass/video/rcmd/Card.kt
@@ -6,7 +6,7 @@ data class Card(
     val adver: Adver,
     val adver_account_id: Int,
     val adver_logo: String,
-    val adver_mid: Int,
+    val adver_mid: Long,
     val adver_name: String,
     val adver_page_url: String,
     val callup_url: String,
