+++ app/src/main/java/cn/spacexc/wearbili/dataclass/videoDetail/Elector.kt
@@ -3,8 +3,8 @@ package cn.spacexc.wearbili.dataclass.videoDetail
 data class Elector(
     val avatar: String,
     val message: String,
-    val mid: Int,
-    val pay_mid: Int,
+    val mid: Long,
+    val pay_mid: Long,
     val rank: Int,
     val trend_type: Int,
     val uname: String,
