+++ app/src/main/java/cn/spacexc/wearbili/dataclass/videoDetail/Live.kt
@@ -2,7 +2,7 @@ package cn.spacexc.wearbili.dataclass.videoDetail
 
 data class Live(
     val endpage_uri: String,
-    val mid: Int,
-    val roomid: Int,
+    val mid: Long,
+    val roomid: Long,
     val uri: String
 )
