+++ app/src/main/java/cn/spacexc/wearbili/dataclass/videoDetail/Reason.kt
@@ -2,7 +2,7 @@ package cn.spacexc.wearbili.dataclass.videoDetail
 
 data class Reason(
     val id: Int,
-    val mid: Int,
+    val mid: Long,
     val name: String,
     val rid: Int,
     val tag_id: Int
